Windows capture provider split into multiple modules to better split responsability
