+++ crates/windows-capture-provider/Cargo.toml
@@ -1,13 +1,11 @@
 [package]
 name = "windows-capture-provider"
-version = "3.0.2"
+version = "4.0.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-scrgb = { workspace = true }
-
 log = { workspace = true }
 thiserror = { workspace = true }
 windows = { workspace = true, features = [
